requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
python-dotenv==1.0.0
Werkzeug==3.0.1
gunicorn==21.2.0
pytest>=8.0,<9.0
langgraph==1.0.7
langchain-openai>=0.3,<1.0
